Update Program.cs
This program was created to demonstrate my understanding of object-oriented programming principles. I focused on encapsulating specific functionality within classes to make the code more readable and maintainable. I also included features that exceed the core requirements, such as the ability to hide words incrementally and the ability to terminate the program under specific conditions.

I hope this code is helpful and informative